pyproject.toml
[tool.poetry]
name = "prompt-security-service"
version = "1.0.0"
description = "Scalable prompt security service"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
redis = "^5.0.0"
alembic = "^1.12.0"
pydantic = "^2.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.0"
prometheus-client = "^0.18.0"
python-dotenv = "^1.0.0"
openai = "^1.3.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"
networkx = "^3.2.0"
python-Levenshtein = "^0.23.0"
pandas = "^2.1.0"
httpx = "^0.25.0"
aiohttp = "^3.9.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pylint = "^3.0.0"
pre-commit = "^3.5.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pylint.messages_control]
disable = "C0111,C0103,R0903"
max-line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"